renderer: Split renderer.h into more source files.
