+++ libopenage/presenter/presenter.cpp
@@ -21,6 +21,8 @@
 #include "renderer/gui/gui.h"
 #include "renderer/gui/integration/public/gui_application_with_logger.h"
 #include "renderer/render_factory.h"
+#include "renderer/render_pass.h"
+#include "renderer/render_target.h"
 #include "renderer/resources/assets/asset_manager.h"
 #include "renderer/resources/shader_source.h"
 #include "renderer/resources/texture_info.h"
@@ -33,6 +35,7 @@
 #include "renderer/window.h"
 #include "time/time_loop.h"
 #include "util/path.h"
+
 
 namespace openage::presenter {
