+++ libopenage/renderer/demo/demo_0.cpp
@@ -4,9 +4,12 @@
 
 #include "renderer/gui/integration/public/gui_application_with_logger.h"
 #include "renderer/opengl/window.h"
+#include "renderer/render_pass.h"
+#include "renderer/render_target.h"
 #include "renderer/resources/mesh_data.h"
 #include "renderer/resources/shader_source.h"
 #include "renderer/shader_program.h"
+
 
 namespace openage::renderer::tests {
