+++ libopenage/renderer/demo/demo_2.cpp
@@ -8,6 +8,8 @@
 
 #include "renderer/gui/integration/public/gui_application_with_logger.h"
 #include "renderer/opengl/window.h"
+#include "renderer/render_pass.h"
+#include "renderer/render_target.h"
 #include "renderer/resources/animation/angle_info.h"
 #include "renderer/resources/animation/frame_info.h"
 #include "renderer/resources/parser/parse_sprite.h"
@@ -16,6 +18,7 @@
 #include "renderer/resources/texture_data.h"
 #include "renderer/shader_program.h"
 #include "renderer/texture.h"
+
 
 namespace openage::renderer::tests {
