+++ libopenage/renderer/demo/demo_5.cpp
@@ -9,6 +9,8 @@
 #include "renderer/camera/camera.h"
 #include "renderer/gui/integration/public/gui_application_with_logger.h"
 #include "renderer/opengl/window.h"
+#include "renderer/render_pass.h"
+#include "renderer/render_target.h"
 #include "renderer/resources/buffer_info.h"
 #include "renderer/resources/shader_source.h"
 #include "renderer/resources/texture_data.h"
