+++ libopenage/renderer/render_target.cpp
@@ -1,0 +1,8 @@
+// Copyright 2024-2024 the openage authors. See copying.md for legal info.
+
+#include "render_target.h"
+
+
+namespace openage::renderer {
+
+} // namespace openage::renderer
