+++ libopenage/renderer/stages/screen/screenshot.cpp
@@ -1,4 +1,4 @@
-// Copyright 2014-2023 the openage authors. See copying.md for legal info.
+// Copyright 2014-2024 the openage authors. See copying.md for legal info.
 
 #include "screenshot.h"
 
@@ -13,7 +13,8 @@
 
 #include "job/job_manager.h"
 #include "log/log.h"
-#include "renderer/renderer.h"
+#include "renderer/render_pass.h"
+#include "renderer/render_target.h"
 #include "renderer/resources/texture_data.h"
 #include "renderer/stages/screen/render_stage.h"
 #include "util/strings.h"
