+++ libopenage/renderer/stages/world/render_stage.cpp
@@ -4,6 +4,8 @@
 
 #include "renderer/camera/camera.h"
 #include "renderer/opengl/context.h"
+#include "renderer/render_pass.h"
+#include "renderer/render_target.h"
 #include "renderer/resources/assets/asset_manager.h"
 #include "renderer/resources/shader_source.h"
 #include "renderer/resources/texture_info.h"
