+++ libopenage/renderer/vulkan/render_target.h
@@ -9,7 +9,7 @@
 
 #include "log/log.h"
 
-#include "renderer/renderer.h"
+#include "renderer/render_target.h"
 #include "renderer/vulkan/graphics_device.h"
